refactor(sidebar): rename `sidebar-overlay` to `sidebar`
